Rationalize OG and meta-description tags
Move default Open Graph tags from base to index so as not to have
vague metadata for non-index pages (e.g., the categories page).

Except for the og:image tag, which we still want to have a
default for in all pages. This is not really valid according to
the OG spec which requires four tags as a minimum, but it should
be beneficial.